Move fake_sound_service
