+++ test/test_container.dart
@@ -14,7 +14,7 @@ import 'package:lichess_mobile/src/model/common/service/sound_service.dart';
 import 'package:lichess_mobile/src/model/auth/auth_repository.dart';
 import 'package:lichess_mobile/src/model/auth/user_session.dart';
 import 'package:lichess_mobile/src/model/auth/session_storage.dart';
-import './common/fake_sound_service.dart';
+import './model/common/service/fake_sound_service.dart';
 import './model/auth/fake_auth_repository.dart';
 import './model/auth/fake_session_storage.dart';
 import './fake_crashlytics.dart';
